+++ src/main/resources/db/migration/V4__Cria_Table_Products.sql
@@ -1,7 +1,7 @@
 CREATE TABLE products (
     id BINARY(16) PRIMARY KEY,
     name VARCHAR(256) NOT NULL,
-    price DECIMAL(10, 2) NOT NULL,
+    price INTEGER NOT NULL,
     category SMALLINT NOT NULL,
     description VARCHAR(256) NOT NULL,
     active BOOLEAN NOT NULL,
